requirements.txt
google_api_python_client==2.36.0
google_auth_oauthlib==0.4.4
requests==2.25.1
